Fix error showing holds list for Evergreen 1.6 The ajax sequence for showing holds list for Evergreen 2.0 cannot work for 1.6, because the opensrf service 'circ.hold.details.retrieve.authoritative' in 1.6 is too narrow in its permission check. We need to use the service if we want to avoid database replication errors, but the service call requires the VIEW_HOLD permission for patrons. However, the permission should not be given because it enables patrons to view the holds of other users.
EG 2.0 fixed this by allowing patrons without the permssion to use the service call for its own holds.
Currently, we do not want to backport the fix to 1.6.

Instead, if we are in a 1.6 context,
we will use the same sequence of ajax calls as the full OPAC,
which will have database replication error,
but at least it will show the holds list.
To determine the context,
we try calling with zero hold id.
If we succeed, we would be in 2.0 context and could show holds list without db replication error;
if we fail, we would use the 1.6 ajax sequence to show holds list.

There is a final problem:
If circ.hold.details.retrieve.authoritative fails,
an ILSEVENT 5000 will occur and the user will normally see an error message.
We disable the notice by using a quick hack to never show any ILSEVENT 5000 errors.
In a more finalized version, we should remove this hack.