Towards a fix for Issue 57: Refactor the login sequence
We move some of the low-level logic from the login_window module to the service
call auth.session.create. It now performs five ajax calls, some in parallel. As
they proceed, the auth.session object gets built incrementally from the server
reponses. The one tricky detail is to remember to handle the server response if
auth.session.complete fails because of bad credentials.