Towards a fix Issue 56: Consider displaying title of overdue item in fines list
1. The service call for getting fines details of an item is fully fieldmapped;
previously we were only fieldmapping the mbts object.  We also normalize the
fieldmap so that it will always have circ and mvr objects, even though for
non-circulation fines they will not exist.

2. In the opac/fines module, we show the item's title and author in the note
field if the transaction type for the fine is 'circulation'; previously we were
only showing the billing note, which is not informative for these types of
fines.