Fix Issue 65: Query string with 'ol' parameter may specify library using numeric ID instead of alphabetic shortname
There are two spots in the software in which the ol parameter is used.

1. In the ou_tree module, we create a helper function to calculate the org unit
ID from ol, whether the ol value is a shortname or already an org unit ID.

2. In the hold_details module, we use a helper function defined in the ou tree
module to calculate the org unit depth from ol.

We also remove an unnecessary dependency for the template module from
hold_details.

In the process, we export the two new helper functions so that other areas of
the software can call them.