Further fix Issue 42: Repackage opac/edit_holds.coffee
We replace the use of jQuery's data method for storing data objects and values
across opac plugins with two alternate mechanisms.

1. To replace the need to store ou tree and ou type, we define a set of helper
functions in the ou_tree module. The functions rely on copies of the objects
stored within the module that have already been requested from the server.
Consequently, the holdings details and hold details modules, which uses the
helper functions, do not have to make redundant server requests for the same
objects.

2. To replace the need to store values related to title details, we pass the
values into the plugins as parameters, making the dependencies more
transparent.