Make the holding_details module more modular
The main level of content for the module was being defined by the edit_hold
module and not by the module itself.  The consequence was that if the module is
not used, some content would still be shown.  Here, we correct this, enabling
us (in a future change) to correctly show content depending on whether the
title has holdings or not.