Fix Issue 70: Version 1.0.1 is broken because search result and title details show missing copy location and copy availability
The error is due to a change in the services API for the
'search.biblio.copy_location_counts.summary.retrieve' request. We modify the
eg/api module to account for the change, which is that call number prefix and
suffix is also included in the result array with the call number label.

We measure the length of the result array and if it equals the lower limit, we
process as before, otherwise we concatenate the call number prefix, label, and
suffix together separated by a space. Typically, the prefix and suffix will be
null values and so we apply jQuery's trim function to the resulting text to
remove leading and trailing white space.

Another related service call has the same change and we apply the same
modification to handle its result.