Better fix for Issue 70
The previous fix did not take into account a further change in the services API
for 'search.asset.retrieve_by_cn', which now needs to be passed an array
consisting of callnumber prefix, label, and suffix, whereas before it needed to
be passed just the label text string.

We now pass the callnumber array from
'search.biblio.copy_location_counts.summary.retrieve' and its relative to the
application software and let the software concatenate them into something
displayable. When we need to make a request for retrieve_by_cn, we have the
input parameter for the callnumber available and in the correct three-component
format.

There is a snag in which we need to decide whether we are in EG version 2.2 or
earlier, because we have to supply the paramter for retrieve_by_cn in the
corresponding format. We use a hack by remembering the format of the callnumber
result from 'search.biblio.copy_location_counts.summary.retrieve' and using
that information to decide how to format the callnumber for retrueve_by_cn.