Fix Issue 71: Page scroll of search result list within title details page is broken
The issue came about after 1.0.1, in which an attempt was made to fix Issue 67:
'Modify title details page when title is an electronic resource'. In that fix,
we tried to determine whether a title had any copies in the search result
module and to publish the relevant topics. However, determining the number of
copies cannot be guaranteed to be done on itme, because of asynchronous service
calls.

In this fix, we go back to having the holding details module determine the
number of copies and have it publish (or not) on a topic that the holds details
module subscribes to.