Fix Issue 77: Cannot see fines summary bar for a user with zero fines
This is a hack to detour around the problem, which seems to be a bug with the
output of actor.user.fines.summary on EG 2.2.  The erroneous output is a null
object instead of data payload containing the mous object, when there are zero
fines.  More puzzling is the error shows up only for a patron account that has
been somehow modified or newly created on the EG 2.2 system. For a patron
account migrated from EG 1.6, there is no error.

To detour the problem, we detect when the output is a null object and build a
suitable data payload ourselves. Part of the payload includes the patron ID,
but we simplify by using zero value, since the software does not need it.