Fix Issue 82: Search bar select menu options can be ordered differently across web browser
Notably Google Chrome seems to show the problem. To fix, we redefine the
settings object for the search bar to be an array of objects, with each object
defining the name and value of one setting. When we loop through the array, we
process each element in the order that they are defined.