Remove focussing logic on main page
By 'focssing logic', we mean what element is auto-focussed upon showing the
main page for the first and subsequent times.  It's complicated because the
main page is composed of several widgets; how do we enforce the priority of one
focus over another?  The current logic is flawed because it is incomplete and
not well coded.  We're commenting it out for now.