Fix Issue 79, Error using IE8
As recommended by jQuery Mobile documention, we need to start processing upon
the document pageinit event and not the document ready event.  When we do this,
we do not get the IE8 error in which the opac tries to focus on an element that
cannot be focussed, namely div.#main, which is the container of the primary
page.