Do not have requirejs modules depend on jQuery or jQueryMobile
1. We need to insert a customization script between the loading of jQ
and jQM.  We cannot achieve this specific loading sequence if jQ and jQM
are participating as requirejs modules.

2. It seems conceptually neater if modules do not depend on jQ or jQM.
Modules can assume jQ or jQM are preloaded and just reference their
global objects.

3. Remove unneeded paths in app.build file.  The file specifies the
build configuration for requirejs r tool.  Now that we do not involve
jQuery and jQueryMobile as requirejs modules, we don't need these paths.