Resequence loading of jQuery and jQueryMobile
According to jQM documentation, it is necessary for mobilesettings to
execute before jQM loads.  Therefore, jQuery is loaded first, followed
by mobilesettings.js, which customizes some global setings for jQM, and
then by jQM.

For the new sequence to work, we cannot have jQ and jQm particpate as
requirejs modules.